Scripts updated
Fixed some bugs, now the disks follow the mouse correctly (I am using old unity control system with raycast to move disks only in X and Y).
Everything is debugged on console in the right way.

We are using a stack to memorize all the disk impiled into a tower, next step is to implement the block if you put a bigger disk onto a smaller one, and the possibility to start (choosing by the start menu) with the first tower with all the disks in the right size order.