Added the rule of "you can't place a disk on to a smaller one"
Added the rule, met a lot of problems, the method ResetToInitialPosition() was somehow bugged with the tower 1 (the middle one) so I've had to add another method called ResetToOrigin() that spawns the disk to tower 0 if that was his initial position.

With this commit I think i finished the logic part of the game and all the scripts needed to make the game work, now I have to implement the UI and a starting menu, later then I will improve the textures of the models and maybe implement some VFX when the disk are stacked and when the game is won.